Avoid NPE when retaining in the view too early
Now we throw our own IllegalStateException with a helpful message if we try to retain something in a view too early. We do this for two reasons: be consistent with the activity and fragment implementations, and to avoid the messageless NPE we would get otherwise.

We could do the same for the other calls that find the saved state registry owner and lifecycle owner, but pragmatically speaking we don’t really have to since it should always fail first when trying to get the viewmodel store, that’s why I decided to keep the double bangs for the other methods calls.